update requirements file and dependency install module
